(refactor): extract code for backstage passes from GildedRose to BackstagePassesUpdater class
